refactor: to remove circular dependencies
