+++ test/tables.reddit.test.ts
@@ -1,10 +1,10 @@
 import {
   hasDieNumber,
-  hasWeights,
   isRedditCollection,
   parseRedditCollection,
   parseWeightedTable,
 } from '../src/module/table/reddit';
+import { hasWeights } from '../src/module/table/lineManipulators';
 
 describe('hasDieNumber', () => {
   it('should return true for a string with a die number', () => {
